+++ packages/breadcrumbs-spaces-react/AvBreadcrumbsSpaces.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { AvBreadcrumbsSpaces } from './';
+import { AvBreadcrumbsSpaces } from '.';
 
 describe('AvBreadcrumbsSpaces', () => {
   test('should render with attributes', () => {