+++ packages/docs/.storybook/addons.js
@@ -1,8 +1,9 @@
-import '@storybook/addon-options/register';
+// Import order determines the order of which the addons appear in the tabs.
+// We want the readme to always show first
 import 'storybook-readme/register';
-
-// Order is important for addons. The order determines which addon shows first in the tab group. We want
-// the readme addon to show first.
 import '@storybook/addon-knobs/register';
-import '@storybook/addon-a11y/register';
 import '@storybook/addon-storysource/register';
+import '@storybook/addon-options/register';
+import '@storybook/addon-a11y/register';
+import '@storybook/addon-actions/register';
+import '@storybook/addon-links/register';