+++ packages/docs/stories/reactstrap-validation-date.stories.js
@@ -3,34 +3,27 @@ import { storiesOf } from '@storybook/react';
 import { withReadme } from 'storybook-readme';
 import { Label, Button } from 'reactstrap';
 import { AvGroup, AvFeedback } from 'availity-reactstrap-validation';
-import {
-  withKnobs,
-  text,
-  boolean,
-  selectV2,
-} from '@storybook/addon-knobs/react';
-import AvFormResults from './mocks/AvFormResults';
-
+import { text, boolean, select } from '@storybook/addon-knobs';
 import AvDate, {
   AvDateField,
   AvDateRange,
   AvDateRangeField,
 } from '@availity/reactstrap-validation-date';
 import README from '@availity/reactstrap-validation-date/README.md';
+import AvFormResults from './mocks/AvFormResults';
 
 const types = {
   text: 'text',
   date: 'date',
 };
-storiesOf('reactstrap Validation|AvDate', module)
+storiesOf('Reactstrap Validation|AvDate', module)
   .addDecorator(withReadme([README]))
-  .addDecorator(withKnobs)
   .add('default', () => (
     <AvFormResults>
       <AvDate
         min={text('Min Date (yyyy-mm-dd)')}
         max={text('Max Date (yyyy-mm-dd)')}
-        type={selectV2('Type', types, 'text')}
+        type={select('Type', types, 'text')}
         name="standAlone"
         aria-label="stand-alone"
         required={boolean('Required', false)}
@@ -42,14 +35,14 @@ storiesOf('reactstrap Validation|AvDate', module)
       </Button>
     </AvFormResults>
   ))
-  .add('default (with own label)', () => (
+  .add('with label', () => (
     <AvFormResults>
       <AvGroup>
         <Label for="standAloneWithLabel">{text('Label', 'AvDate Label')}</Label>
         <AvDate
           min={text('Min Date (yyyy-mm-dd)')}
           max={text('Max Date (yyyy-mm-dd)')}
-          type={selectV2('Type', types, 'text')}
+          type={select('Type', types, 'text')}
           name="standAlone"
           aria-label="stand-alone"
           required={boolean('Required', false)}
@@ -71,7 +64,7 @@ storiesOf('reactstrap Validation|AvDate', module)
           label={text('Label', 'Field Label')}
           min={text('Min Date (yyyy-mm-dd)')}
           max={text('Max Date (yyyy-mm-dd)')}
-          type={selectV2('Type', types, 'text')}
+          type={select('Type', types, 'text')}
           name="AvDateField"
           required={boolean('Required', false)}
           disabled={boolean('Disabled', false)}
@@ -96,7 +89,7 @@ storiesOf('reactstrap Validation|AvDate', module)
       <AvFormResults>
         <AvDateRange
           name="AvDateRange"
-          type={selectV2('Type', types, 'text')}
+          type={select('Type', types, 'text')}
           start={{ name: 'date.start' }}
           end={{ name: 'date.end' }}
           min={text('Min Date (yyyy-mm-dd)')}
@@ -124,7 +117,7 @@ storiesOf('reactstrap Validation|AvDate', module)
         <AvDateRangeField
           label={text('Label', 'Date Range')}
           name="AvDateRange"
-          type={selectV2('Type', types, 'text')}
+          type={select('Type', types, 'text')}
           start={{ name: 'date.start', required }}
           end={{ name: 'date.end', required }}
           min={text('Min Date (yyyy-mm-dd)')}