+++ packages/docs/stories/upload.stories.js
@@ -1,14 +1,8 @@
 import React, { Component, createRef } from 'react';
 import { storiesOf } from '@storybook/react';
 import { withReadme } from 'storybook-readme';
-import {
-  withKnobs,
-  button,
-  boolean,
-  number,
-  array,
-} from '@storybook/addon-knobs/react';
-import mock, { proxy } from 'xhr-mock';
+import { button, boolean, number, array } from '@storybook/addon-knobs';
+import mock from 'xhr-mock';
 import { Button, CustomInput } from 'reactstrap';
 import { AvForm, AvField, AvGroup } from 'availity-reactstrap-validation';
 import UploadCore from '@availity/upload-core';
@@ -19,6 +13,7 @@ import Upload, {
   FilePicker,
   FilePickerBtn,
 } from '@availity/upload';
+
 import README from '@availity/upload/README.md';
 import MockUpload from '@availity/upload/tests/mockUpload';
 
@@ -86,8 +81,6 @@ class PreUpload extends Component {
   }
 }
 
-mock.setup();
-
 mock.post(
   /\/ms\/api\/availity\/internal\/core\/vault\/upload\/v1\/resumable\/[^/]\//,
   (req, res) =>
@@ -139,13 +132,12 @@ mock.use(
       .body('')
 );
 
-mock.use(proxy);
+// mock.use(proxy);
 
 let instance;
 
-storiesOf('Actions|Upload', module)
+storiesOf('Components|Upload', module)
   .addDecorator(withReadme([README]))
-  .addDecorator(withKnobs)
   .addDecorator(story => {
     instance = new MockUpload();
     return story();
@@ -216,7 +208,5 @@ storiesOf('Actions|Upload', module)
         />
       </div>
     );
-  });
-storiesOf('Actions|Upload/With Metadata', module).add('Pre-upload', () => (
-  <PreUpload />
-));
+  })
+  .add('Pre-upload', () => <PreUpload />);