+++ packages/upload/tests/FilePicker.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { FilePicker } from '../';
+import { FilePicker } from '..';
 
 describe('Upload', () => {
   test('should render', () => {