+++ packages/upload/tests/mockUpload.js
@@ -5,12 +5,19 @@ export default class MockUpload {
     id += 1;
     return id.toString();
   })();
+
   percentage = 0;
+
   errorMessage = null;
+
   status = 'pending';
+
   onProgress = [];
+
   onSuccess = [];
+
   onError = [];
+
   sendPassword = password => {
     if (password) {
       this.success();
@@ -18,6 +25,7 @@ export default class MockUpload {
       this.error('Incorrect password', 'encrypted');
     }
   };
+
   progress = (amount = 10) => {
     this.percentage = Math.min(Math.max(amount + this.percentage, 0), 100);
     this.errorMessage = null;
@@ -26,17 +34,20 @@ export default class MockUpload {
       this.onSuccess.forEach(a => a());
     }
   };
+
   success = () => {
     this.percentage = 100;
     this.errorMessage = null;
     this.status = 'accepted';
     this.onSuccess.forEach(a => a());
   };
+
   error = (message = null, status = 'rejected') => {
     this.errorMessage = message;
     this.status = status;
     this.onError.forEach(a => a());
   };
+
   reset = () => {
     this.percentage = 0;
     this.errorMessage = null;